gwegew

+++ YouFeel/Module/Applications/Statistics/EquipmentStatus/EquipmentAvailability/Views/BXTEPSummaryView.m
@@ -109,7 +109,7 @@
 
 - (IBAction)btnClick:(UIButton *)sender
 {
-    NSLog(@"----------------- %ld", sender.tag);
+    NSLog(@"----------------- %ld", (long)sender.tag);
     BXTEquipmentListViewController *listVC = [[BXTEquipmentListViewController alloc] init];
     listVC.date = self.timeStr;
